added init manifesto
this should be a sister project of todos 521, they will likely be combined into a singluar project as modular components - that for planning to-dos, and this for documenting any actions, and a third for putting things into a schedule and common daily interface. And finally how can these (and any and all other apps) be customized through 1 interface (that every individual can use to customize all software to their personal wokring style).